feat: network id refactor

+++ crates/game-client/src/networking/messages.rs
@@ -1,5 +1,5 @@
 use bevy::prelude::*;
-use game_core::components::NetworkId;
+use game_core::networking::NetworkId;
 
 /// Internal message for combat hit feedback, emitted by network event handling.
 #[derive(Message)]
@@ -16,23 +16,23 @@ pub struct ActorDespawnMessage(pub NetworkId);
 
 #[derive(Message)]
 pub struct SpellImpactMessage {
-    pub target_id: u64,
+    pub target_id: u32,
     pub spell_id: u32,
     pub impact_amount: i32,
 }
 
 #[derive(Message)]
-pub struct ActorDeathMessage(pub u64);
+pub struct ActorDeathMessage(pub u32);
 
 #[derive(Message)]
 pub struct StartCastingMessage {
-    pub actor_id: u64,
+    pub actor_id: u32,
     pub spell_id: u32,
 }
 
 #[derive(Message)]
 pub struct KillRewardMessage {
-    pub victim_id: u64,
+    pub victim_id: u32,
     pub loot: Vec<protocol::models::ItemDrop>,
 }
